+++ packages/console_test_app/bin/sqflite_ffi_simple_bin.dart
@@ -1,4 +1,3 @@
-import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 Future<void> main(List<String> arguments) async {